requirements.txt
fastapi
uvicorn
jinja2
python-multipart
pymongo
transformers
torch
numpy
pydantic
python-dotenv